backend/requirements.txt
Flask==2.3.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.0
web3==6.11.0
py-solc-x==2.0.4
pytest==7.4.0
pytest-flask==1.2.0
PyJWT==2.8.0
bcrypt==4.1.1
gunicorn==21.2.0
waitress==2.1.2